Add WaitSet::new_for_node()
This function is important for users who directly use wait sets, since the live_subscriptions() etc. functions are not public.
